Create FileHolder to normalize NCSD
NCSD was constructed with a std::fs::File while NCCH and RomFS received
a read::Reader. To make these all the same, and to clean up the setup
code in the bin/ files, I created a FileHolder::open(filename) that took
care of all the initialization code (getting the file length and
creating a RwLock).