Add RomFS::file_at and fix a lot of lifetime issues
I didn't quite understand the lifetime of my objects, but
while implementing romfs::RomFS::file_at I hit a problem where going
down the directory tree was binding each object to its parent, so I had
a bunch of temporary values being dropped and causing problems.

I fixed this by making explicitly marking the lifetime of the return
value to be the same as the Reader<'a> and unbinding them from the
object (no more &'a self). When I added the &'a selfs I didn't quite
understand what I was doing.